NavBar conversion complete
